ci/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via
    #   dagger-io
    #   httpcore
attrs==23.1.0
    # via
    #   cattrs
    #   dagger-io
backoff==2.2.1
    # via gql
beartype==0.14.0
    # via dagger-io
cattrs==22.2.0
    # via dagger-io
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via typer
colorama==0.4.6
    # via typer
dagger-io==0.5.3
    # via -r requirements.in
gql==3.4.1
    # via dagger-io
graphql-core==3.2.3
    # via
    #   dagger-io
    #   gql
h11==0.14.0
    # via httpcore
httpcore==0.17.0
    # via httpx
httpx==0.24.0
    # via dagger-io
idna==3.4
    # via
    #   anyio
    #   httpx
    #   yarl
markdown-it-py==2.2.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via yarl
platformdirs==3.5.1
    # via dagger-io
pygments==2.15.1
    # via rich
rich==13.3.5
    # via
    #   dagger-io
    #   typer
shellingham==1.5.0.post1
    # via typer
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typer[all]==0.9.0
    # via dagger-io
typing-extensions==4.5.0
    # via
    #   dagger-io
    #   typer
yarl==1.9.2
    # via gql
